+++ src/mesido/financial_mixin.py
@@ -144,6 +144,19 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
                 )
                 nominal_variable_operational = nominal_fixed_operational
                 nominal_investment = nominal_fixed_operational
+            elif asset_name in [*self.energy_system_components.get("cold_demand", [])]:
+                nominal_fixed_operational = (
+                    bounds[f"{asset_name}.Cold_demand"][1]
+                    if not np.isinf(bounds[f"{asset_name}.Cold_demand"][1])
+                    else bounds[f"{asset_name}.HeatIn.Heat"][1]  # TODO: This In or Out to be confirmed?
+                )
+                nominal_fixed_operational = (
+                    nominal_fixed_operational
+                    if isinstance(nominal_fixed_operational, float)
+                    else max(nominal_fixed_operational.values)
+                )
+                nominal_variable_operational = nominal_fixed_operational
+                nominal_investment = nominal_fixed_operational
             elif asset_name in [*self.energy_system_components.get("heat_source", [])]:
                 nominal_fixed_operational = self.variable_nominal(f"{asset_name}.Heat_source")
                 nominal_fixed_operational = (
@@ -432,6 +445,7 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
         for asset in [
             *self.energy_system_components.get("heat_source", []),
             *self.energy_system_components.get("heat_demand", []),
+            *self.energy_system_components.get("cold_demand", []),
             *self.energy_system_components.get("ates", []),
             *self.energy_system_components.get("low_temperature_ates", []),
             *self.energy_system_components.get("heat_buffer", []),
@@ -458,6 +472,7 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             for asset in [
                 *self.energy_system_components.get("heat_source", []),
                 *self.energy_system_components.get("heat_demand", []),
+                *self.energy_system_components.get("cold_demand", []),
                 *self.energy_system_components.get("ates", []),
                 *self.energy_system_components.get("low_temperature_ates", []),
                 *self.energy_system_components.get("heat_buffer", []),
@@ -875,6 +890,8 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
 
         parameters = self.parameters(ensemble_member)
 
+        timesteps_hr = np.diff(self.times()) / 3600
+
         for asset in [
             *self.energy_system_components.get("ates", []),
             *self.energy_system_components.get("low_temperature_ates", []),
@@ -903,11 +920,9 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             else:
                 price_profile = Timeseries(self.times(), np.zeros(len(self.times())))
 
-            timesteps = np.diff(self.times()) / 3600.0
-
             sum = 0.0
             for i in range(1, len(self.times())):
-                sum += price_profile.values[i] * pump_power[i] * timesteps[i - 1] / eff
+                sum += price_profile.values[i] * pump_power[i] * timesteps_hr[i - 1] / eff
 
             constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
 
@@ -921,7 +936,6 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             variable_operational_cost_coefficient = parameters[
                 f"{s}.variable_operational_cost_coefficient"
             ]
-            timesteps = np.diff(self.times()) / 3600.0
 
             pump_power = self.__state_vector_scaled(f"{s}.Pump_power", ensemble_member)
             eff = parameters[f"{s}.pump_efficiency"]
@@ -951,10 +965,24 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
                 sum += (
                     variable_operational_cost_coefficient
                     * heat_source[i]
-                    * timesteps[i - 1]
+                    * timesteps_hr[i - 1]
                     / denominator
                 )
-                sum += price_profile.values[i] * pump_power[i] * timesteps[i - 1] / eff
+                sum += price_profile.values[i] * pump_power[i] * timesteps_hr[i - 1] / eff  # pump_power ?
+
+                # TODO: resolve issue for elect costs fir elec heat pump (ie. heat source), see code below
+                # if (
+                #     s not in self.energy_system_components.get("air_water_heat_pump", [])
+                #     and s not in self.energy_system_components.get("air_water_heat_pump_elec", [])
+                # ):
+                #     sum += price_profile.values[i] * pump_power[i] * timesteps_hr[i - 1] / eff
+                # else:
+                #     sum += (
+                #         price_profile.values[i]
+                #         * heat_source[i]
+                #         * timesteps_hr[i - 1]
+                #         / denominator
+                #     )
 
             constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
 
@@ -970,7 +998,6 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             variable_operational_cost_coefficient = parameters[
                 f"{hp}.variable_operational_cost_coefficient"
             ]
-            timesteps = np.diff(self.times()) / 3600
             pump_power = self.__state_vector_scaled(f"{hp}.Pump_power", ensemble_member)
             eff = parameters[f"{hp}.pump_efficiency"]
 
@@ -989,13 +1016,31 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             sum = 0.0
             for i in range(1, len(self.times())):
                 sum += (
-                    variable_operational_cost_coefficient * elec_consumption[i] * timesteps[i - 1]
+                    variable_operational_cost_coefficient
+                    * elec_consumption[i]
+                    * timesteps_hr[i - 1]
                 )
-                sum += price_profile.values[i] * pump_power[i] * timesteps[i - 1] / eff
+                sum += price_profile.values[i] * pump_power[i] * timesteps_hr[i - 1] / eff
                 if hp not in self.energy_system_components.get("heat_pump_elec", []):
                     # assuming that if heatpump has electricity port, the cost for the electricity
                     # are already made by the electricity producer and transport
-                    sum += price_profile.values[i] * elec_consumption[i] * timesteps[i - 1]
+                    sum += price_profile.values[i] * elec_consumption[i] * timesteps_hr[i - 1]
+
+            constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
+
+        for ac in self.energy_system_components.get("airco", []):
+            heat_airco = self.__state_vector_scaled(f"{ac}.Heat_airco", ensemble_member)
+            variable_operational_cost_var = self._asset_variable_operational_cost_map[ac]
+            variable_operational_cost = self.extra_variable(
+                variable_operational_cost_var, ensemble_member
+            )
+            nominal = self.variable_nominal(variable_operational_cost_var)
+            variable_operational_cost_coefficient = parameters[
+                f"{ac}.variable_operational_cost_coefficient"
+            ]
+            sum = 0.0
+            for i in range(1, len(self.times())):
+                sum += variable_operational_cost_coefficient * heat_airco[i] * timesteps_hr[i - 1]
 
             constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
 
@@ -1017,9 +1062,12 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             ]
 
             sum = 0.0
-            timesteps = np.diff(self.times()) / 3600.0
             for i in range(1, len(self.times())):
-                sum += variable_operational_cost_coefficient * gas_mass_flow[i] * timesteps[i - 1]
+                sum += (
+                    variable_operational_cost_coefficient
+                    * gas_mass_flow[i]
+                    * timesteps_hr[i - 1]
+                )
 
             constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
 
@@ -1035,11 +1083,13 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             variable_operational_cost_coefficient = parameters[  # euro / g
                 f"{gs}.variable_operational_cost_coefficient"
             ]
-            timesteps = np.diff(self.times())
+            timesteps_sec = np.diff(self.times())
             sum = 0.0
             for i in range(1, len(self.times())):
                 sum += (
-                    variable_operational_cost_coefficient * gas_produced_g_s[i] * timesteps[i - 1]
+                    variable_operational_cost_coefficient
+                    * gas_produced_g_s[i]
+                    * timesteps_sec[i - 1]
                 )  # [euro/g] * [g/s] * [s]
             constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
 
@@ -1055,11 +1105,10 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             variable_operational_cost_coefficient = parameters[  # euro / Wh
                 f"{es}.variable_operational_cost_coefficient"
             ]
-            timesteps = np.diff(self.times()) / 3600.0  # convert dt from [s] to [hr]
             sum = 0.0
             for i in range(1, len(self.times())):
                 sum += (
-                    variable_operational_cost_coefficient * elec_produced_w[i] * timesteps[i - 1]
+                    variable_operational_cost_coefficient * elec_produced_w[i] * timesteps_hr[i - 1]
                 )  # [euro/Wh] * [W] * [hr]
             constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))
 
@@ -1075,13 +1124,12 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
         # variable_operational_cost_coefficient = parameters[
         #     f"{a}.variable_operational_cost_coefficient"
         # ]
-        # timesteps = np.diff(self.times()) / 3600.0
         #
         # sum = 0.0
         #
         # for i in range(1, len(self.times())):
         #     varOPEX_dt = (variable_operational_cost_coefficient * heat_ates[i]
-        #     * timesteps[i - 1])
+        #     * timesteps_hr[i - 1])
         #     constraints.append(((varOPEX-varOPEX_dt)/nominal,0.0, np,inf))
         #     #varOPEX would be a variable>0 for everyt timestep
         #     sum += varOPEX
@@ -1102,12 +1150,11 @@ class FinancialMixin(BaseComponentTypeMixin, CollocatedIntegratedOptimizationPro
             ]
 
             sum = 0.0
-            timesteps = np.diff(self.times()) / 3600.0
             for i in range(1, len(self.times())):
                 sum += (
                     variable_operational_cost_coefficient
                     * power_consumer[i]
-                    * timesteps[i - 1]  # gas_mass_flow unit is g/s
+                    * timesteps_hr[i - 1]  # gas_mass_flow unit is g/s
                 )
 
             constraints.append(((variable_operational_cost - sum) / nominal, 0.0, 0.0))